common/src/main/java/jdev/dto/Point.java
package jdev.dto;

import com.fasterxml.jackson.core.JsonProcessingException;
import com.fasterxml.jackson.databind.ObjectMapper;

/**
 * Created by artem on 26.12.23.
 */
public class Point {
    private double latitude; //широта
    private double longitude; //долгота
    private double azimuth; //
    private double speed; //

    public double getLatitude() {
        return latitude;
    }

    public void setLatitude(double latitude) {
        this.latitude = latitude;
    }

    public double getLongitude() {
        return longitude;
    }

    public void setLongitude(double longitude) {
        this.longitude = longitude;
    }

    public double getAzimuth() {
        return azimuth;
    }

    public void setAzimuth(double azimuth) {
        this.azimuth = azimuth;
    }

    public double getSpeed() {
        return speed;
    }

    public void setSpeed(double speed) {
        this.speed = speed;
    }



/*    public String toJson_old() {
        return "Point{" +
                "latitude=" + latitude +
                ", longitude=" + longitude +
                ", autoId='" + autoId + '\'' +
                '}';
    }*/

    public String toJson() throws JsonProcessingException {
        ObjectMapper mapper = new ObjectMapper();
        return mapper.writeValueAsString(this);
    }

    @Override
    public String toString() {
        return "Point{" +
                "latitude=" + latitude +
                ", longitude='" + longitude + '\'' +
                '}';
    }

}
